summary method for projection plots.

summary method for projection plots.

# S3 method for class 'optproj'
summary(object, xnames, ...)

Arguments

object

An optproj object, i.e., output from the function optim_proj().

xnames

Optional vector of names for the elements of the potential solution.

...

Further arguments to be passed to or from other methods.

Value

An object of class summary.optproj inheriting from summary.optcheck, with elements:

xsol

The potential solution vector.

ysol

The value of the objective function at xsol.

maximize

Logical indicating whether the potential solution should maximize or minimize the objective function.

xopt

A vector containing the argmax/argmin in each projection plot.

yopt

A vector containing the max/min in each projection plot.

xdiff

A two-column matrix containing the differences between xsol and xopt. The first column is the absolute difference D = xopt - xsol, the second is the relative difference R = D/|xsol|.

ydiff

Same thing, but between ysol and yopt.

Details

The print methods for summary.optproj and optproj objects themselves both return a three-column matrix, consisting of the potential solution (xsol), the optimal solution in each projection plot (xopt), and the relative difference between the two (R = (xopt - xsol)/|xsol|).

See also